PR comment 1/n: Move registerMBean to utils
